# openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "EmpireIdle API",
    "description": "Browser-based idle empire builder game API",
    "version": "v1"
  },
  "paths": {
...
    "/api/market/{playerId}": {
      "get": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_Browse",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MarketListingKind"
            }
          },
          {
            "name": "itemKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MarketPageView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketPageView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketPageView"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{playerId}/mine": {
      "get": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_Mine",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MyMarketView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMarketView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMarketView"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{playerId}/quote": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_Quote",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "\u0423\u043D\u0456\u043A\u0430\u043B\u044C\u043D\u0438\u0439 \u043A\u043B\u044E\u0447 \u043E\u043F\u0435\u0440\u0430\u0446\u0456\u0457 \u2014 \u0437\u0430\u0445\u0438\u0441\u0442 \u0432\u0456\u0434 \u043F\u043E\u0434\u0432\u0456\u0439\u043D\u043E\u0433\u043E \u0432\u0438\u043A\u043E\u043D\u0430\u043D\u043D\u044F \u043F\u0440\u0438 \u0440\u0435\u0442\u0440\u0430\u044F\u0445",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketGoodsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketGoodsRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/MarketGoodsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MarketQuoteView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketQuoteView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketQuoteView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{playerId}/listings": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_List",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "\u0423\u043D\u0456\u043A\u0430\u043B\u044C\u043D\u0438\u0439 \u043A\u043B\u044E\u0447 \u043E\u043F\u0435\u0440\u0430\u0446\u0456\u0457 \u2014 \u0437\u0430\u0445\u0438\u0441\u0442 \u0432\u0456\u0434 \u043F\u043E\u0434\u0432\u0456\u0439\u043D\u043E\u0433\u043E \u0432\u0438\u043A\u043E\u043D\u0430\u043D\u043D\u044F \u043F\u0440\u0438 \u0440\u0435\u0442\u0440\u0430\u044F\u0445",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListOnMarketRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListOnMarketRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ListOnMarketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MarketListingCreated"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketListingCreated"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketListingCreated"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{playerId}/listings/{listingId}/buy": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_Buy",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "\u0423\u043D\u0456\u043A\u0430\u043B\u044C\u043D\u0438\u0439 \u043A\u043B\u044E\u0447 \u043E\u043F\u0435\u0440\u0430\u0446\u0456\u0457 \u2014 \u0437\u0430\u0445\u0438\u0441\u0442 \u0432\u0456\u0434 \u043F\u043E\u0434\u0432\u0456\u0439\u043D\u043E\u0433\u043E \u0432\u0438\u043A\u043E\u043D\u0430\u043D\u043D\u044F \u043F\u0440\u0438 \u0440\u0435\u0442\u0440\u0430\u044F\u0445",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{playerId}/listings/{listingId}/cancel": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_Cancel",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "\u0423\u043D\u0456\u043A\u0430\u043B\u044C\u043D\u0438\u0439 \u043A\u043B\u044E\u0447 \u043E\u043F\u0435\u0440\u0430\u0446\u0456\u0457 \u2014 \u0437\u0430\u0445\u0438\u0441\u0442 \u0432\u0456\u0434 \u043F\u043E\u0434\u0432\u0456\u0439\u043D\u043E\u0433\u043E \u0432\u0438\u043A\u043E\u043D\u0430\u043D\u043D\u044F \u043F\u0440\u0438 \u0440\u0435\u0442\u0440\u0430\u044F\u0445",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListOnMarketRequest": {
        "required": [
          "kind",
          "priceGold",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/MarketListingKind"
          },
          "equipmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "heroId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "itemKey": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "priceGold": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "MarketEquipmentView": {
        "required": [
          "enhancementLevel",
          "rarity",
          "slot",
          "stats"
        ],
        "type": "object",
        "properties": {
          "slot": {
            "type": "string"
          },
          "rarity": {
            "type": "string"
          },
          "enhancementLevel": {
            "type": "integer",
            "format": "int32"
          },
          "stats": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "additionalProperties": false
      },
      "MarketGoodsRequest": {
        "required": [
          "kind",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/MarketListingKind"
          },
          "equipmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "heroId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "itemKey": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MarketHeroView": {
        "required": [
          "constellation",
          "level",
          "tier"
        ],
        "type": "object",
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "tier": {
            "type": "integer",
            "format": "int32"
          },
          "constellation": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MarketListingCreated": {
        "required": [
          "listingId"
        ],
        "type": "object",
        "properties": {
          "listingId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "MarketListingKind": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "MarketListingView": {
        "required": [
          "expiresAt",
          "id",
          "isOwn",
          "itemKey",
          "kind",
          "listedAt",
          "priceGold",
          "pricePerUnit",
          "quantity",
          "state",
          "units"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "itemKey": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "priceGold": {
            "type": "integer",
            "format": "int32"
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          },
          "units": {
            "type": "number",
            "format": "double"
          },
          "listedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "state": {
            "type": "string"
          },
          "isOwn": {
            "type": "boolean"
          },
          "equipment": {
            "$ref": "#/components/schemas/MarketEquipmentView"
          },
          "hero": {
            "$ref": "#/components/schemas/MarketHeroView"
          }
        },
        "additionalProperties": false
      },
      "MarketPageView": {
        "required": [
          "listings",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "listings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketListingView"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MarketQuoteView": {
        "required": [
          "maxPrice",
          "minPrice",
          "suggestedPrice",
          "taxShare",
          "units"
        ],
        "type": "object",
        "properties": {
          "units": {
            "type": "number",
            "format": "double"
          },
          "minPrice": {
            "type": "integer",
            "format": "int32"
          },
          "maxPrice": {
            "type": "integer",
            "format": "int32"
          },
          "suggestedPrice": {
            "type": "integer",
            "format": "int32"
          },
          "taxShare": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "MyMarketView": {
        "required": [
          "activeListings",
          "isOpen",
          "listingHours",
          "listingLimit",
          "listings",
          "taxShare"
        ],
        "type": "object",
        "properties": {
          "isOpen": {
            "type": "boolean"
          },
          "opensAtTownHall": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "listingLimit": {
            "type": "integer",
            "format": "int32"
          },
          "activeListings": {
            "type": "integer",
            "format": "int32"
          },
          "taxShare": {
            "type": "number",
            "format": "double"
          },
          "listingHours": {
            "type": "integer",
            "format": "int32"
          },
          "listings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketListingView"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "\u0412\u0432\u0435\u0434\u0438 JWT \u0442\u043E\u043A\u0435\u043D (\u0431\u0435\u0437 \u0441\u043B\u043E\u0432\u0430 Bearer)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}