# src/EmpireIdle.Domain/Enums/MarketListingKind.cs
namespace EmpireIdle.Domain.Enums
{
    /// <summary>Що виставлено на ринок.</summary>
    public enum MarketListingKind
    {
        /// <summary>Екземпляр спорядження зі своїми статами й заточкою.</summary>
        Equipment = 1,

        /// <summary>Герой без спорядження.</summary>
        Hero = 2,

        /// <summary>Пачка стакового предмета.</summary>
        Item = 3
    }
}